feat: Initial testing of alternative llm framework - pocketflow
Pocketflow is a minimalist 100-line llm framework library to help with creating llm applications
Its core principles involve creating nodes and flows for designing these systems
In these tests I tried to create an instance of a "Pocketflow Parsing Agent" as a flow between two nodes:
the first is a LoadJS node, which retrieves script information from a site
the second is a ReadJS node, which sends the script information to a long-context llm (gemini-2.0-flash-exp) to read and identify any useful information

the information between the two nodes can be shared through a storage dictionary "shared" and information can be retrieved or written to it with these nodes

Overall, it seems like a useful strategy if creaeting tools/agents with simpler functionalities in view. Larger/more complex frameworks seem to complicate the agent execution process.